+++ modules/client/src/internal/client/methods.ts
@@ -4,12 +4,14 @@ import {
   DAOFactory__factory,
   DAORegistry__factory,
   PluginRepo__factory,
+  PluginSetupProcessor__factory,
 } from "@aragon/core-contracts-ethers";
 import {
   GraphQLError,
   InvalidAddressOrEnsError,
   NoProviderError,
   NoSignerError,
+  MissingExecPermissionError
 } from "@aragon/sdk-common";
 import { Signer } from "@ethersproject/abstract-signer";
 import { BigNumber } from "@ethersproject/bignumber";
@@ -41,6 +43,7 @@ import {
   IDepositParams,
   IHasPermissionParams,
   ITransferQueryParams,
+  PermissionIds,
   SubgraphBalance,
   SubgraphDao,
   SubgraphDaoListItem,
@@ -49,11 +52,7 @@ import {
   Transfer,
   TransferSortBy,
 } from "../../interfaces";
-import {
-  ClientCore,
-  Context,
-  SortDirection,
-} from "../../client-common";
+import { ClientCore, Context, SortDirection } from "../../client-common";
 import {
   toAssetBalance,
   toDaoDetails,
@@ -82,7 +81,7 @@ export class ClientMethods extends ClientCore implements IClientMethods {
    * @memberof ClientMethods
    */
   public async *create(
-    params: ICreateParams,
+    params: ICreateParams
   ): AsyncGenerator<DaoCreationStepValue> {
     const signer = this.web3.getConnectedSigner();
     if (!signer) {
@@ -93,14 +92,14 @@ export class ClientMethods extends ClientCore implements IClientMethods {
 
     const daoFactoryInstance = DAOFactory__factory.connect(
       this.web3.getDaoFactoryAddress(),
-      signer,
+      signer
     );
 
     const pluginInstallationData: DAOFactory.PluginSettingsStruct[] = [];
     for (const plugin of params.plugins) {
       const latestVersion = await PluginRepo__factory.connect(
         plugin.id,
-        signer,
+        signer
       ).getLatestVersion();
       pluginInstallationData.push({
         pluginSetup: latestVersion[1],
@@ -116,13 +115,46 @@ export class ClientMethods extends ClientCore implements IClientMethods {
       throw new Error("Could not pin the metadata on IPFS");
     }
 
+    // check if at least one plugin requests EXECUTE_PERMISSION on the DAO
+    // This check isn't 100% correct all the time
+    // simulate the DAO creation to get an address
+    const pluginSetupProcessorAddr = await daoFactoryInstance.pluginSetupProcessor();
+    const pluginSetupProcessor = PluginSetupProcessor__factory.connect(
+      pluginSetupProcessorAddr,
+      signer
+    );
+    let execPermissionFound = false;
+
+    // using the DAO base because it reflects a newly created DAO the best
+    const daoBaseAddr = await daoFactoryInstance.daoBase()
+    // simulates each plugin installation seperately to get the requested permissions
+    for (const plugin of pluginInstallationData) {
+      const pluginSetupProcessorResponse = await pluginSetupProcessor.callStatic.prepareInstallation(
+        daoBaseAddr, 
+        plugin.pluginSetup,
+        plugin.pluginSetupRepo,
+        plugin.data
+      );
+      const found = pluginSetupProcessorResponse.permissions.find(
+        permission => permission.permissionId === PermissionIds.EXECUTE_PERMISSION_ID
+      );
+      if (found) {
+        execPermissionFound = true;
+        break;
+      }
+    }
+
+    if(!execPermissionFound) {
+      throw new MissingExecPermissionError()
+    }
+
     const tx = await daoFactoryInstance.connect(signer).createDao(
       {
         name: params.ensSubdomain,
         metadata: toUtf8Bytes(cid),
         trustedForwarder: params.trustedForwarder || AddressZero,
       },
-      pluginInstallationData,
+      pluginInstallationData
     );
 
     yield {
@@ -134,9 +166,9 @@ export class ClientMethods extends ClientCore implements IClientMethods {
     const daoFactoryInterface = DAORegistry__factory.createInterface();
     // find dao address using the dao registry address
     const log = receipt.logs?.find(
-      (e) =>
+      e =>
         e.topics[0] ===
-          id(daoFactoryInterface.getEvent("DAORegistered").format("sighash")),
+        id(daoFactoryInterface.getEvent("DAORegistered").format("sighash"))
     );
     if (!log) {
       throw new Error("Failed to create DAO");
@@ -161,7 +193,7 @@ export class ClientMethods extends ClientCore implements IClientMethods {
    * @memberof ClientMethods
    */
   public async *deposit(
-    params: IDepositParams,
+    params: IDepositParams
   ): AsyncGenerator<DaoDepositStepValue> {
     const signer = this.web3.getConnectedSigner();
     if (!signer) {
@@ -170,7 +202,7 @@ export class ClientMethods extends ClientCore implements IClientMethods {
       throw new Error("A web3 provider is needed");
     }
     const [daoAddress, amount, tokenAddress, reference] = unwrapDepositParams(
-      params,
+      params
     );
 
     if (tokenAddress && tokenAddress !== AddressZero) {
@@ -180,7 +212,7 @@ export class ClientMethods extends ClientCore implements IClientMethods {
         daoAddress,
         amount.toBigInt(),
         tokenAddress,
-        signer,
+        signer
       );
     }
 
@@ -197,20 +229,20 @@ export class ClientMethods extends ClientCore implements IClientMethods {
       tokenAddress,
       amount,
       reference,
-      override,
+      override
     );
     yield { key: DaoDepositSteps.DEPOSITING, txHash: depositTx.hash };
 
-    await depositTx.wait().then((cr) => {
+    await depositTx.wait().then(cr => {
       if (!cr.logs?.length) {
         throw new Error("The deposit was not properly registered");
       }
 
       const daoInterface = DAO__factory.createInterface();
       const log = cr.logs?.find(
-        (e) =>
+        e =>
           e?.topics[0] ===
-            id(daoInterface.getEvent("Deposited").format("sighash")),
+          id(daoInterface.getEvent("Deposited").format("sighash"))
       );
       if (!log) {
         throw new Error("Failed to deposit");
@@ -219,11 +251,9 @@ export class ClientMethods extends ClientCore implements IClientMethods {
       const logParsed = daoInterface.parseLog(log);
       if (!amount.eq(logParsed.args["amount"])) {
         throw new Error(
-          `Deposited amount mismatch. Expected: ${amount.toBigInt()}, received: ${
-            logParsed.args[
-              "amount"
-            ].toBigInt()
-          }`,
+          `Deposited amount mismatch. Expected: ${amount.toBigInt()}, received: ${logParsed.args[
+            "amount"
+          ].toBigInt()}`
         );
       }
     });
@@ -234,13 +264,13 @@ export class ClientMethods extends ClientCore implements IClientMethods {
     daoAddress: string,
     amount: bigint,
     tokenAddress: string,
-    signer: Signer,
+    signer: Signer
   ): AsyncGenerator<DaoDepositStepValue> {
     const tokenInstance = new Contract(tokenAddress, erc20ContractAbi, signer);
 
     const currentAllowance = await tokenInstance.allowance(
       await signer.getAddress(),
-      daoAddress,
+      daoAddress
     );
 
     yield {
@@ -252,7 +282,7 @@ export class ClientMethods extends ClientCore implements IClientMethods {
 
     const tx: ContractTransaction = await tokenInstance.approve(
       daoAddress,
-      BigNumber.from(amount),
+      BigNumber.from(amount)
     );
 
     yield {
@@ -262,7 +292,7 @@ export class ClientMethods extends ClientCore implements IClientMethods {
 
     await tx.wait().then((cr: ContractReceipt) => {
       const log = cr.logs?.find(
-        (e) => e?.topics[0] === id("Approval(address,address,uint256)"),
+        e => e?.topics[0] === id("Approval(address,address,uint256)")
       );
       if (!log) {
         throw new Error("Could not find the Approval event");
@@ -285,9 +315,7 @@ export class ClientMethods extends ClientCore implements IClientMethods {
    * @return {*}  {Promise<boolean>}
    * @memberof ClientMethods
    */
-  public async hasPermission(
-    params: IHasPermissionParams,
-  ): Promise<boolean> {
+  public async hasPermission(params: IHasPermissionParams): Promise<boolean> {
     const signer = this.web3.getConnectedSigner();
     if (!signer) {
       throw new NoSignerError();
@@ -295,15 +323,12 @@ export class ClientMethods extends ClientCore implements IClientMethods {
       throw new NoProviderError();
     }
     // connect to the managing dao
-    const daoInstance = DAO__factory.connect(
-      params.daoAddressOrEns,
-      signer,
-    );
+    const daoInstance = DAO__factory.connect(params.daoAddressOrEns, signer);
     return daoInstance.hasPermission(
       params.where,
       params.who,
       id(params.permission),
-      params.data || new Uint8Array([]),
+      params.data || new Uint8Array([])
     );
   }
   /**
@@ -341,7 +366,7 @@ export class ClientMethods extends ClientCore implements IClientMethods {
       // this is a temporal fix and should be changed by the line above
       // but the current daos in subgraph dont have a proper metadata
       const stringMetadata = await this.ipfs.fetchString(
-        "QmebY8BGzWBUyVqZFMaFkFmz3JrfaDoFP5orDqzJ1uiEkr",
+        "QmebY8BGzWBUyVqZFMaFkFmz3JrfaDoFP5orDqzJ1uiEkr"
       );
       const metadata = JSON.parse(stringMetadata);
       return toDaoDetails(dao, metadata);
@@ -377,7 +402,7 @@ export class ClientMethods extends ClientCore implements IClientMethods {
           skip,
           direction,
           sortBy,
-        },
+        }
       );
       await this.ipfs.ensureOnline();
       return Promise.all(
@@ -388,12 +413,12 @@ export class ClientMethods extends ClientCore implements IClientMethods {
             // this is a temporal fix and should be changed by the line above
             // but the current daos in subgraph dont have a proper metadata
             const test_cid = "QmebY8BGzWBUyVqZFMaFkFmz3JrfaDoFP5orDqzJ1uiEkr";
-            return this.ipfs.fetchString(test_cid).then((stringMetadata) => {
+            return this.ipfs.fetchString(test_cid).then(stringMetadata => {
               const metadata = JSON.parse(stringMetadata);
               return toDaoListItem(dao, metadata);
             });
-          },
-        ),
+          }
+        )
       );
     } catch (err) {
       throw new GraphQLError("DAO");
@@ -409,7 +434,7 @@ export class ClientMethods extends ClientCore implements IClientMethods {
    */
   public async getBalances(
     daoAddressorEns: string,
-    _tokenAddresses: string[],
+    _tokenAddresses: string[]
   ): Promise<AssetBalance[] | null> {
     let address = daoAddressorEns;
     if (!isAddress(address)) {
@@ -439,8 +464,8 @@ export class ClientMethods extends ClientCore implements IClientMethods {
       // handle other tokens that are not ERC20 or eth
       return Promise.all(
         balances.map(
-          (balance: SubgraphBalance): AssetBalance => toAssetBalance(balance),
-        ),
+          (balance: SubgraphBalance): AssetBalance => toAssetBalance(balance)
+        )
       );
     } catch (err) {
       throw new GraphQLError("balance");
@@ -501,16 +526,15 @@ export class ClientMethods extends ClientCore implements IClientMethods {
           skip,
           direction,
           sortBy,
-        },
+        }
       );
       if (!vaultTransfers) {
         return null;
       }
       return Promise.all(
         vaultTransfers.map(
-          (transfer: SubgraphTransferListItem): Transfer =>
-            toTransfer(transfer),
-        ),
+          (transfer: SubgraphTransferListItem): Transfer => toTransfer(transfer)
+        )
       );
     } catch {
       throw new GraphQLError("transfer");