+++ modules/client/CHANGELOG.md
@@ -16,6 +16,8 @@ TEMPLATE:
 - Fix 1, 2, 3
 -->
 ## [UPCOMING]
+### Added
+- Checks if at least one plugin requests `EXECUTE_PERMISSION` during DAO creation
 
 ## 0.14.0-alpha
 On 2022-11-22 16:45:52